listening for changes in language
Successfully listening for changes in language using the settings, but still have to add the logic for transmitting  the language name further as a parameter for Blockly to be able to change its language as well.

Setting another language in Blockly, doesn't change the names of the categories to the respective language.